impr(fe): markdown code block pretty rendering
